Refactor to have cli directory.
