+++ bin/tower-cli
@@ -14,72 +14,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import importlib
-import os
-import sys
-
 import click
 
-import tower_cli
-from tower_cli import __version__
-from tower_cli.utils import secho
-
-
-class TowerCLI(click.MultiCommand):
-    """Tower CLI is a command-line interface tool for interacting with
-    [Ansible Tower][1]. It allows basic CRUD operations and job control
-    from the Unix shell.
-
-      [1]: http://www.ansible.com/tower/
-    """
-    def list_commands(self, ctx):
-        """Return a list of commands present in the commands and resources
-        folders, but not subcommands.
-        """
-        answer = set()
-        locations = ('commands', 'resources')
-        for loc in locations:
-            path = '%s/%s/' % (tower_cli.whereami, loc)
-            for filename in os.listdir(path):
-                if filename.endswith('.py') and not filename.startswith('_'):
-                    if loc == 'resources':
-                        res = tower_cli.get_resource(filename[:-3])
-                        if not getattr(res, 'internal', False):
-                            answer.add(filename[:-3])
-                    else:
-                        answer.add(filename[:-3])
-        return sorted(answer)
-
-    def get_command(self, ctx, name):
-        """Given a command identified by its name, import the appropriate
-        module and return the decorated command.
-
-        Resources are automatically commands, but if both a resource and
-        a command are defined, the command takes precedence.
-        """
-        # First, attempt to get a basic command from `tower_cli.commands`.
-        try:
-            module = importlib.import_module('tower_cli.commands.%s' % name)
-            return getattr(module, name)
-        except ImportError:
-            pass
-
-        # No command was found; try to get a resource.
-        try:
-            resource = tower_cli.get_resource(name)
-            return resource.as_command()
-        except ImportError:
-            pass
-
-        # Okay, we weren't able to find a command.        
-        secho('No such command: %s.' % name, fg='red', bold=True)
-        sys.exit(2)
-
-    def invoke(self, ctx):
-        if ctx.params.get('version', False):
-            click.echo('Tower CLI %s' % __version__)
-        else:
-            return super(TowerCLI, self).invoke(ctx)
+from tower_cli.cli.base import TowerCLI
 
 
 if __name__ == '__main__':
