+++ tests/test_resources_ad_hoc.py
@@ -16,6 +16,7 @@
 import tower_cli
 from tower_cli.api import client
 from tower_cli.utils import exceptions as exc
+from tower_cli.cli.resource import ResSubcommand
 
 from tests.compat import unittest, mock
 from tower_cli.conf import settings
@@ -83,7 +84,7 @@ class LaunchTests(unittest.TestCase):
             self.assertEqual(result['changed'], True)
             self.assertEqual(result['id'], 42)
 
-            f = self.res.as_command()._echo_method(self.res.launch)
+            f = ResSubcommand(self.res)._echo_method(self.res.launch)
             with mock.patch.object(click, 'secho'):
                 with settings.runtime_values(format='human'):
                     f(inventory="foobar", machine_credential=2)
