+++ tests/test_resources_workflow.py
@@ -85,7 +85,7 @@ class NodeModelTests(unittest.TestCase):
         """
         Test application of additional decorator in __getattribute__
         """
-        with mock.patch('tower_cli.models.base.ResourceMethods.write') as mck:
+        with mock.patch('tower_cli.models.base.BaseResource.write') as mck:
             mck.return_value = {'id': 589}
             mck.__name__ = 'create'
             self.res.create(workflow_job_template=1,
